backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
openai>=1.0.0
tiktoken>=0.4.0
pandas>=2.0.0
requests>=2.28.2
beautifulsoup4>=4.11.2
lxml>=4.9.2
python-multipart>=0.0.6
httpx>=0.24.0
asyncpg>=0.27.0
uuid>=1.30
pytest>=7.3.1
httpx>=0.24.0
python-jose[cryptography]>=3.3.0
aiofiles>=23.1.0